+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/QuotedAttachmentFactory.kt
@@ -32,6 +32,7 @@ import io.getstream.chat.android.uiutils.extension.hasLink
  * for the quoted message.
  */
 public object QuotedAttachmentFactory : AttachmentFactory(
+    type = Type.BuiltIn.QUOTED,
     canHandle = {
         it.firstOrNull()
             ?.let { attachment ->