+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/composer/AudioRecordingController.kt
@@ -157,7 +157,7 @@ internal class AudioRecordingController(
         }
     }
 
-    public fun startRecording() {
+    public fun startRecording(offset: Pair<Float, Float>? = null) {
         val state = this.recordingState.value
         if (state !is RecordingState.Idle) {
             logger.w { "[startRecording] rejected (state is not Idle): $state" }
@@ -166,7 +166,21 @@ internal class AudioRecordingController(
         logger.i { "[startRecording] state: $state" }
         val recordingName = "audio_recording_${Date()}"
         mediaRecorder.startAudioRecording(recordingName, realPollingInterval.toLong())
-        this.recordingState.value = RecordingState.Hold()
+        this.recordingState.value = RecordingState.Hold(offset = offset ?: RecordingState.Hold.ZeroOffset)
+    }
+
+    public fun holdRecording(offset: Pair<Float, Float>? = null) {
+        val state = this.recordingState.value
+        if (state !is RecordingState.Hold) {
+            logger.w { "[holdRecording] rejected (state is not Hold): $state" }
+            return
+        }
+        if (offset == null) {
+            logger.v { "[holdRecording] rejected (offset is null)" }
+            return
+        }
+        logger.v { "[holdRecording] offset: Offset(${offset.first}:${offset.second})" }
+        this.recordingState.value = state.copy(offset = offset)
     }
 
     public fun lockRecording() {
@@ -187,6 +201,9 @@ internal class AudioRecordingController(
         }
         logger.i { "[cancelRecording] state: $state" }
         mediaRecorder.release()
+        if (state is RecordingState.Overview && state.isPlaying) {
+            audioPlayer.resetAudio(state.playingId)
+        }
         clearData()
         this.recordingState.value = RecordingState.Idle
     }
@@ -461,6 +478,7 @@ internal class AudioRecordingController(
     }
 
     private fun List<Int>.downsampleToSingleMax(): Int {
+        1 to 1
         return max()
     }
 }