+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/application/ChatInitializer.kt
@@ -190,5 +190,13 @@ class ChatInitializer(
         }*/
 
         ChatUI.decoratorProviderFactory = CustomDecoratorProviderFactory() + DecoratorProviderFactory.defaultFactory()
+
+        TransformStyle.messageComposerStyleTransformer = StyleTransformer { defaultStyle ->
+            defaultStyle.copy(
+                audioRecordingButtonVisible = true,
+                audioRecordingButtonEnabled = true,
+                audioRecordingButtonPreferred = true,
+            )
+        }
     }
 }