+++ stream-chat-android-ui-utils/src/main/kotlin/io/getstream/chat/android/uiutils/extension/Attachment.kt
@@ -18,6 +18,7 @@ package io.getstream.chat.android.uiutils.extension
 
 import io.getstream.chat.android.client.extensions.uploadId
 import io.getstream.chat.android.client.utils.attachment.isAudio
+import io.getstream.chat.android.client.utils.attachment.isAudioRecording
 import io.getstream.chat.android.client.utils.attachment.isFile
 import io.getstream.chat.android.client.utils.attachment.isVideo
 import io.getstream.chat.android.models.Attachment
@@ -30,7 +31,8 @@ public fun Attachment.isAnyFileType(): Boolean {
         upload != null ||
         isFile() ||
         isVideo() ||
-        isAudio()
+        isAudio() ||
+        isAudioRecording()
 }
 
 /**